Spam/requirements_B.txt
tensorflow==2.18.0
fastapi>=0.110,<0.120
uvicorn>=0.29,<0.30
numpy==1.26.0
pillow>=9.0,<11.0
python-multipart>=0.0.6